backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
requests==2.32.3
pydantic==2.8.2
stripe==9.10.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
cachetools==5.3.3
tenacity==8.5.0
